backend-gcloud-run/requirements.txt
flask
flask-cors
firebase-admin
requests
bcrypt
stripe
google-auth
requests
PyJWT
python-dotenv